requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-dotenv
yfinance
numpy
pandas
# psycopg[binary]   # only if using Postgres
